% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prod_igr.R
\name{calc_prod_igr}
\alias{calc_prod_igr}
\title{calc_prod_igr}
\usage{
calc_prod_igr(
  taxaSampleListMass = NULL,
  taxaInfo = NULL,
  bootNum = NULL,
  dateDf = NULL,
  taxaSummary = "full",
  wrap = TRUE,
  massValue = NULL,
  massLabel = NULL,
  ...
)
}
\arguments{
\item{taxaSampleListMass}{description}

\item{taxaInfo}{The taxa info data.frame}

\item{bootNum}{integer. The number of bootstrap samples to produce.}

\item{dateDf}{data frame of date information with external predictors for each month. There should be a column name identical to all variables in the growth equation found in taxaInfo data.frame.}

\item{taxaSummary}{string of 'short', 'full', or 'none'. What type of summary information should be returned.}

\item{wrap}{logical. Should the dates wrap to create a full year?}

\item{massValue}{string. What is the mass value and units of the production}

\item{massLabel}{string. What label should the output units be. It is possible this will default to 'massValue' in the future.}

\item{...}{additional arguments to pass to the function}
}
\value{
returns a list of 2 objects
}
\description{
This function calculates production using the instantaneous growth rate method.
}
\details{
Additional details...
}
