% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prod_is.R
\name{calc_prod_is}
\alias{calc_prod_is}
\title{calc_prod_is}
\usage{
calc_prod_is(
  taxaSampleListMass = NULL,
  taxaInfo = NULL,
  bootNum = NULL,
  dateDf = NULL,
  taxaSummary = "full",
  wrap = FALSE,
  massValue = "afdm_mg",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  bootList = NULL,
  ...
)
}
\arguments{
\item{taxaSampleListMass}{description}

\item{taxaInfo}{data frame of taxonomic information for calculating production}

\item{bootNum}{integer. How many bootstrap samples should be constructed}

\item{dateDf}{data frame of date information with external predictors for each month. There should be a column name identical to all variables in the growth equation found in taxaInfo data.frame.}

\item{taxaSummary}{string of 'short', 'full', or 'none'. What type of summary information should be returned.}

\item{wrap}{logical. Should the dates wrap to create a full year?}

\item{massValue}{string. What is the mass value and units of the production}

\item{bootList}{list. This is the bootstrapped samples passed from `calc_production()`}

\item{...}{additional arguments to be passed to the function}

\item{massLabel}{string. What label should the output units be. It is possible this will default to 'massValue' in the future.}
}
\value{
returns a list of 2 objects:

P.boots: the boostrapped estimates of production, abundance, and biomass.

taxaSummary: is the summary of the sample production, abundance, and biomass
}
\description{
This function calculates secondary production with the increment-summation method.
}
