% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prod_sf.R
\name{calc_prod_sf}
\alias{calc_prod_sf}
\title{calc_prod_sf}
\usage{
calc_prod_sf(
  taxaSampleListMass = NULL,
  taxaInfo = NULL,
  bootNum = NULL,
  taxaSummary = TRUE,
  wrap = FALSE,
  lengthValue = NULL,
  massValue = "mass",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  bootList = NULL,
  ...
)
}
\arguments{
\item{taxaSampleListMass}{description}

\item{taxaInfo}{data frame of taxonomic information for calculating production}

\item{bootNum}{integer. How many bootstrap samples should be constructed}

\item{taxaSummary}{logical. If TRUE (default) the taxaSummary will be with annual summary information will be returned.}

\item{wrap}{logical. Should the dates wrap to create a full year?}

\item{lengthValue}{string of the column name containing the length class measurements}

\item{massValue}{string of the column name containing the mass measurement}

\item{abunValue}{string of the column name containing the abundance or density measurement}

\item{dateCol}{string of the column name containing the date information. This can be either a recognized date object (e.g., Date, POSIX) or numeric.}

\item{repCol}{string of the column name containing the replicate information}

\item{bootList}{list. This is the bootstrapped samples passed from \code{calc_production()}}

\item{...}{additional arguments to be passed to the function}
}
\value{
returns a list of 2 objects:

P.boots: the boostrapped estimates of production, abundance, and biomass.

taxaSummary: is the summary of the sample production, abundance, and biomass
}
\description{
This function calculates secondary production with the size-frequency method.
}
