% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_production.R
\name{calc_production}
\alias{calc_production}
\title{calc_production}
\usage{
calc_production(
  taxaSampleListMass = NULL,
  taxaInfo = NULL,
  bootNum = 100,
  taxaSummary = TRUE,
  lengthValue = NULL,
  massValue = "mass",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  envData = NULL,
  ...
)
}
\arguments{
\item{taxaSampleListMass}{data.frame of sample length-masses and abundances}

\item{taxaInfo}{dataframe (or coercible); The taxonomic information for calculating secondary production. This must include a taxonomic ID column with the same name as that of \code{taxaSampleListMass}}

\item{bootNum}{integer. The number of bootstrapped samples to create}

\item{taxaSummary}{logical. If TRUE (default) the taxaSummary will be with annual summary information will be returned.}

\item{lengthValue}{string of the column name containing length class measurements. If NULL (default)}

\item{massValue}{string of the column name containing the mass measurement}

\item{abunValue}{string of the column name containing the abundance or density measurement}

\item{dateCol}{string of the column name containing the date information. This can be either a recognized date object (e.g., Date, POSIX)}

\item{repCol}{string of the column name containing the replicate information}

\item{envData}{data.frame of additional environmental data for each sampling event.  If present, this is combined with the dateDf information and passed to the IGR function for growth rate estimations.The environmental variables passed should align with the start of each sampling interval.}

\item{...}{additional arguments passed to function}

\item{infoCols}{integer vector; Any columns in the sizeInfo object that are not the taxonomic ID, sampling metadata, or size class columns. ** This is soft-deprecated and is not currently used. It may go away soon **}
}
\description{
This is the main function of the secpRod package. It will calculate secondary production for all groups based on the methods described in the taxa information object. Depending on input values varying summaries are returned.
}
