% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{findreps}
\alias{findreps}
\title{findreps}
\source{
https://stackoverflow.com/questions/33155662/find-and-break-on-repeated-runs
}
\usage{
findreps(x, counter = NULL)
}
\arguments{
\item{x}{character string of the formula structure}

\item{counter}{stuff}
}
\description{
This function finds and sets breaks on repeated runs of similar character types. It is used to parse the formula structure and variables in `massForm` for converting lengths to mass in \code{convert_length_to_mass()}.
}
