% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_prod.sample.R
\name{igr_prod.sample}
\alias{igr_prod.sample}
\title{igr_prod.sample}
\usage{
igr_prod.sample(df = NULL, sizesDf = NULL, cpi = NULL, full = TRUE, ...)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{sizesDf}{a data.frame of the}

\item{cpi}{integer. The cohort production interval.}

\item{full}{logical. should the full summary be returned with mean and sd}

\item{...}{additional arguments passed to function}
}
\value{
list object with taxa summary of the sampled data
}
\description{
This function calculates taxa production based on the instantaneous growth method
}
