% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prod.sample.R
\name{is_prod.sample}
\alias{is_prod.sample}
\title{is_prod.sample}
\usage{
is_prod.sample(
  df = NULL,
  dateDf = dateDf,
  massValue = "afdm_mg",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  wrap = FALSE,
  full = TRUE,
  ...
)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{massValue}{character string identifying the column name of the mass value}

\item{full}{logical. should the full summary be returned with mean and sd}

\item{...}{additional arguments passed to function}

\item{sizesDf}{a data.frame of the}

\item{massLabel}{character string identifying the column name of the mass value}
}
\value{
list object with taxa summary of the sampled data
}
\description{
This function calculates taxa production based on the increment-summation method
}
