% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_boot.comp.R
\name{my_boot.comp}
\alias{my_boot.comp}
\title{my_boot.comp}
\usage{
my_boot.comp(
  y,
  x = NULL,
  N = NULL,
  max.comp = 2,
  B = 100,
  sig = 0.05,
  arbmean = TRUE,
  arbvar = TRUE,
  mix.type = c("logisregmix", "multmix", "mvnormalmix", "normalmix", "poisregmix",
    "regmix", "regmix.mixed", "repnormmix"),
  hist = TRUE,
  ...
)
}
\arguments{
\item{y}{The raw data for multmix, mvnormalmix, normalmix, and repnormmix and the response values for logisregmix, poisregmix, and regmix. See the documentation concerning their respective EM algorithms for specific structure of the raw data.}

\item{x}{The predictor values required only for the regression mixtures logisregmix, poisregmix, and regmix. A column of 1s for the intercept term must not be included! See the documentation concerning their respective EM algorithms for specific structure of the predictor values.}

\item{N}{An n-vector of number of trials for the logistic regression type logisregmix. If NULL, then N is an n-vector of 1s for binary logistic regression.}

\item{max.comp}{The maximum number of components to test for. The default is 2. This function will perform a test of k-components versus (k+1)-components sequentially until we fail to reject the null hypothesis. This decision rule is governed by the calculated p-value and sig.}

\item{B}{The number of bootstrap realizations of the likelihood ratio statistic to produce. The default is 100, but ideally, values of 1000 or more would be more acceptable.}

\item{sig}{The significance level for which to compare the p-value against when performing the test of k-components versus (k+1)-components.}

\item{arbmean}{If FALSE, then a scale mixture analysis can be performed for mvnormalmix, normalmix, regmix, or repnormmix. The default is TRUE.}

\item{arbvar}{If FALSE, then a location mixture analysis can be performed for mvnormalmix, normalmix, regmix, or repnormmix. The default is TRUE.}

\item{mix.type}{The type of mixture analysis you wish to perform. The data inputted for y and x depend on which type of mixture is selected. logisregmix corresponds to a mixture of logistic regressions. multmix corresponds to a mixture of multinomials with data determined by the cut-point method. mvnormalmix corresponds to a mixture of multivariate normals. normalmix corresponds to a mixture of univariate normals. poisregmix corresponds to a mixture of Poisson regressions. regmix corresponds to a mixture of regressions with normal components. regmix.mixed corresponds to a mixture of regressions with random or mixed effects. repnormmix corresponds to a mixture of normals with repeated measurements.}

\item{hist}{An argument to provide a matrix plot of histograms for the boostrapped likelihood ratio statistic.}

\item{...}{Additional arguments passed to the various EM algorithms for the mixture of interest}
}
\value{
\code{boot.comp} returns a list with items:\item{p.values}{The p-values for each test of k-components versus (k+1)-components.} \item{log.lik}{The B bootstrap realizations of the likelihood ratio statistic.} \item{obs.log.lik}{The observed likelihood ratio statistic for each test which is used in determining the p-values.}
}
\description{
my_boot.comp
}
