% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_prod.sample.R
\name{pb_prod.sample}
\alias{pb_prod.sample}
\title{pb_prod.sample}
\usage{
pb_prod.sample(
  df = NULL,
  sizesDf = NULL,
  massValue = NULL,
  massLabel = NULL,
  pb = NULL,
  full = TRUE,
  ...
)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{sizesDf}{a data.frame of the}

\item{massValue}{character string identifying the column name of the mass value}

\item{massLabel}{character string identifying the column name of the mass value}

\item{pb}{numeric. The production:biomass ratio.}

\item{full}{logical. should the full summary be returned with mean and sd}

\item{...}{additional arguments passed to function}
}
\value{
list object with taxa summary of the sampled data
}
\description{
This function calculates taxa production based on the production:biomass method
}
