% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cohorts.R
\name{plot_cohorts}
\alias{plot_cohorts}
\title{plot_cohorts}
\usage{
plot_cohorts(
  taxaSampleListMass = NULL,
  param = c("length", "mass"),
  massClass = "afdm_mg",
  ...
)
}
\arguments{
\item{taxaSampleListMass}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{param}{character. A string of 'length' or 'mass' that describes what measurement should be plotted}

\item{massClass}{character. The column name of the mass measurement (e.g., afdm_mg, dm_mg, etc.)}

\item{...}{additional arguments passed to function}
}
\value{
returns a histogram of the plot of the relative frequency of size or mass classes for a single taxon for all sampling dates
}
\description{
This function plots size-frequency data of a single taxon over time
}
