% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_boots.R
\name{prep_boots}
\alias{prep_boots}
\title{prep_boots}
\usage{
prep_boots(df = NULL, bootNum = bootNum, dateCol = "dateID", repCol = "repID")
}
\arguments{
\item{df}{data.frame. A dataframe of the species size, mass, and frequency data.}

\item{bootNum}{integer. The number of bootstrapped data sets that should be created.}
}
\description{
This function prepares bootstrap samples.
}
