% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_production.R
\name{production_igr}
\alias{production_igr}
\title{Instanteonous growth rate method production estimates}
\usage{
production_igr(growth, biomass, size.class)
}
\arguments{
\item{growth}{data.frame containing sampling event and size-specific site
instantaneous growth rates. Should contain columns for sample interval (cum),
length, age, interval length (interval).}

\item{biomass}{data.frame containing size scrutruce data with columns for
site, sample interval (cum),length, weight (dry_wt), and area.}

\item{size.class}{numeric value indicating the denominator for size class
estimation For example, 30 indicates size classes are 30 day cohorts}
}
\value{
data.frame summarized at the site/sample interval level containing
original grouping variables and estimates of  biomass, standing biomass, and
production.
}
\description{
Estimates production, standing biomass, and PtoB ratios using instantaneous
growth rate method. Production is estimated for all sites and sampling
intervals contained in a data frame using size/age class specific growth
rates and biomass summarized at the sample interval.

A short description...
}
