% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_production.R
\name{production_igr_boot}
\alias{production_igr_boot}
\title{Bootstrapped instantaneous growth rate production estimates}
\usage{
production_igr_boot(
  growth,
  biomass,
  size.class,
  n.sim,
  return.raw = F,
  parallel = F,
  mc.cores = NULL
)
}
\value{
description
}
\description{
A short description...
}
