% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_cohorts.R
\name{reconstruct_split_cohort}
\alias{reconstruct_split_cohort}
\title{reconstruct_split_cohort}
\usage{
reconstruct_split_cohort(
  df,
  timeCol = "dateID",
  massCol = "massValue",
  massDropThresh = 0.6,
  tStart = 5,
  models = c("vbg", "gompertz", "logistic", "richards"),
  offsetBounds = c(10, 150),
  fallbackGrid = TRUE
)
}
\arguments{
\item{df}{a data.frame of the sample-level observed masses and densities}

\item{timeCol}{character. String of the column name depicting the date of each sampling event.}

\item{massCol}{character. String of the column name containing mass data.}

\item{massDropThresh}{This is the proportional drop in mass used to detect the end and beginning of two cohort portions.}

\item{tStart}{This is the guess for the starting value of age (in days) of the smallest mean size sampled.}

\item{models}{character. String vector of the names of models to fit to estimate \eqn{t_0}. See `details` for more information.}

\item{offsetBounds}{integer vector of length = 2. The lower and upper bounds of the offset to test for fit}

\item{fallbackGrid}{logical. If TRUE (default) a grid search procedure will be used if `optim()` fails.}
}
\description{
This function is used to reconstruct a *non-overlapping* cohort sampled over two years, often with a period of
of zero abundance (e.g. due to egg or adult stages).
}
\details{
The initiation of a sampling program can often start when the growth of a population cohort is already underway, leading to observing only large, late instars individuals during early sampling events.
For univoltine species, this may be followed by a period of zero abundance (e.g., due to egg or adult stages) and subsequent sampling of early instars. To make inferences on cohort parameters
and use some model-based procedures to estimate production, it is useful to 're-order' the sampling dates to create a continuous cohort progression. This makes some assumptions. For more details on these,
see the _sampling simulation vignette_. This function reorders partial cohorts (e.g., late
instars at the beginning and early instars in later sampling in an annual regime. To do this, the model fits growth functions to remap split cohorts to a continuous progression in time. The difficult
aspect of this is estimating the relative cohort age of different sampling events for correct ordering when there may be a long period between. Crucially, we don't observe the beginning or end of a full
cohort, we only have information from our distinct sampling events. This function fits growth models to estimate the cohort ages of each sample and remap into cohort progression.

Models are fit to mean cohort mass at each sampling event, \eqn{M_{t}} and return estimates for asymptotic mass, \eqn{M_\inf}, time at mass = 0, \eqn{t_0} (or time at growth inflection, \eqn{t*},
depending on model which is used to estimate \eqn{t_0}), and growth rate, \eqn{k}. These  Current models available are:

     "vbg" The 'regular' von Bertalanffy growth model: \deqn{M_{t} = M_{\inf} (1 -e^{-k \dot (t - t_0)})}
     "gompertz" The Gompertz growth model: \deqn{M_t = M_\inf e^{e^{-k \dot (t - t*)}}}
     "logistic" The logistic growth model: \deqn{M_t = M_\inf / (1 + e^{-k \dot (t - t*)})}
     "richards" The Richards growth model: \deqn{M_t = M_\inf (1 + 1/D e^{-k \dot (t - t*)})^{-D}}

These
}
