% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_cohorts.R
\name{fit_with_offset}
\alias{fit_with_offset}
\alias{plot_cohort_fit}
\title{fit_with_offset}
\usage{
fit_with_offset(
  dfOrdered,
  offset,
  models = c("vbg", "gompertz", "logistic", "richards"),
  tStart = 5
)

plot_cohort_fit(remappedCohort, models = "ensemble", labelPoints = TRUE)
}
\arguments{
\item{dfOrdered}{the reordered sampling data set from [reconstruct_split_cohort()] processes}

\item{offset}{the offset (in days) between the final sample and first sample in two cohort portions to be joined}

\item{models}{character. String vector of the names of models to fit to \eqn{M_t}. See `details` for more information.}

\item{tStart}{data frame of date information with external predictors for each month. There should be a column name identical to all variables in the growth equation found in taxaInfo data.frame.}

\item{remappedCohort}{the reordered object returned from [reconstruct_split_cohort()]}

\item{labelPoints}{logical. Should the points be labelled with the sampling date information}
}
\value{
returns a list of three (3) objects:

fits: the growth model fits to reordered sampling dates

aiccs: the Akaike Information Criterion for each model. This is used to build ensemble model estimates for remapped cohort 'times'

df_pseudo: a data.frame of the original date, standardized cohort 'times', and t0 corrected ages estimated from the growth models.

returns a ggplot object of the
}
\description{
This is an internal function used in [reconstruct_split_cohort()] to fit growth functions to find the optimal cohort offset in a split cohort.

This is an internal function used in [reconstruct_split_cohort()] to fit growth functions to find the optimal cohort offset in a split cohort.
}
\seealso{
[reconstruct_split_cohorts()] for documentation on the process used to reconstruct split cohorts

[reconstruct_split_cohorts()] for documentation on the process used to reconstruct split cohorts
}
