% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_interval.R
\name{sample_interval}
\alias{sample_interval}
\title{Calculate the amount of time between two sampling periods.}
\usage{
sample_interval(df)
}
\arguments{
\item{df}{data.frame containing columns for sample site (site), sample period}
}
\value{
data.frame with columns for site, sample period, date, and sample
}
\description{
The time between two user indicated sampling periods is
calculated and assigned to the first sampling period in each pair. For use
in interval growth estimation.
}
