% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_prod.sample.R
\name{sf_prod.sample}
\alias{sf_prod.sample}
\title{sf_prod.sample}
\usage{
sf_prod.sample(
  df = NULL,
  sizesDf = NULL,
  lengthValue = NULL,
  massValue = massValue,
  abunValue = abunValue,
  dateCol = dateCol,
  repCol = repCol,
  wrap = FALSE,
  cpi = NULL,
  full = TRUE,
  ...
)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{sizesDf}{a data.frame of the size class including lengthClass, massClass, bin_min, bin_max, midpoint}

\item{lengthValue}{string of the column name containing the length class measurements}

\item{massValue}{character string identifying the column name of the mass value}

\item{abunValue}{character string identifying the column name of the density value}

\item{dateCol}{character string identifying the column name of the sample date information. This is distinguished from *Value parameters in that Values may be used to maintain units provenance in the future. This may also change.}

\item{repCol}{character string identifying the column name of the replicate information.}

\item{wrap}{logical. should the calculations be wrapped by adding an additional date to make a full year?}

\item{cpi}{integer. The cohort production interval.}

\item{full}{logical. should the full summary be returned with mean and sd}

\item{...}{additional arguments passed to function}
}
\value{
list object with annual production, mean biomass, and mean abundance of the sample
}
\description{
This function calculates taxa production based on the size-frequency method
}
