% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singleCohortSim}
\alias{singleCohortSim}
\title{Sample data for community secondary production analysis}
\format{
`singleCohortSim`
A single tibble data frame with a list-col simulating the sampling of a single cohort population:
\describe{
  \item{x}{integer. x location of sampled grid cell}
  \item{y}{integer. y location of sampled grid cell}
  \item{larvalDensity}{integer. The density of larvae in the grid cell}
  \item{massDistribution}{numeric. list-col of the mass of each individual in the cell}
}
}
\source{
data-raw/single_cohort_sim.R
}
\usage{
singleCohortSim
}
\description{
A collection of community macroinvertebrate data for estimating secondary production
}
\keyword{datasets}
