% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{univoltine}
\alias{univoltine}
\title{Sample data for community secondary production analysis}
\format{
`univoltine`
A single data.frame of the sampleInfo data as sampled for a single simulated taxon with columns:
\describe{
  \item{taxonID}{the taxonomic identifier}
  \item{repID}{the replicate identifier}
  \item{dateID}{the date identifier as.Date format}
  \item{lengthClass}{the numeric (coercible) description of individual length}
  \item{n_m2}{the count or density of individuals from a lengthClass}
}
}
\usage{
univoltine
}
\description{
A collection of community macroinvertebrate data for estimating secondary production
}
\keyword{datasets}
