% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wbtData}
\alias{wbtData}
\title{Sample data for community secondary production analysis}
\format{
`wbtData`
A list of two objects. 'sampleInfo' is list of 32 data frames for each taxonomic entity and the second, 'taxaInfo' is a data frame with 32 rows and 13 columns:
\describe{
  \item{taxonID}{the taxonomic identifier}
  \item{repID}{the replicate identifier}
  \item{dateID}{the date identifier as.Date format}
  \item{lengthClass}{the numeric (coercible) description of individual length}
  \item{n_m2}{the count or density of individuals from a lengthClass}
}
}
\source{
<https://doi.org/10.4319/lo.2014.59.2.0507>
}
\usage{
wbtData
}
\description{
A collection of community macroinvertebrate data for estimating secondary production
}
\keyword{datasets}
