% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_dates.R
\name{wrap_dates}
\alias{wrap_dates}
\title{wrap_dates}
\usage{
wrap_dates(df = NULL, envData = NULL, dateCol = NULL, wrapDate = TRUE)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{envData}{(optional) a data.frame of the environmental data. The number of rows should be the same as the number of sampling dates.}

\item{dateCol}{character string identifying the column name of the sampling dates}

\item{wrapDate}{logical. should we wrap the dates to make full year?}
}
\value{
data.frame with sampling dates and intervals, or number of days between
}
\description{
This function creates an additional date in the data set to make a full annual cycle
}
