% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_production.R
\name{calc_production}
\alias{calc_production}
\title{calc_production}
\usage{
calc_production(
  taxaSampleListMass = NULL,
  taxaInfo = NULL,
  bootNum = 100,
  taxaSummary = "full",
  lengthValue = NULL,
  massValue = "afdm_mg",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  ...
)
}
\arguments{
\item{taxaSampleListMass}{data.frame of sample length-masses and abundances}

\item{taxaInfo}{dataframe (or coercible); The taxonomic information for calculating secondary production. This must include a taxonomic ID column with the same name as that of \code{taxaSampleListMass}}

\item{bootNum}{integer. The number of bootstrapped samples to create}

\item{taxaSummary}{string of \code{'short'}, \code{'full'}, or \code{'none'} to distinguish the information returned}

\item{lengthValue}{string of the column name containing length class measurements. If NULL (default)}

\item{massValue}{string of the column name containing the mass measurement}

\item{abunValue}{string of the column name containing the abundance or density measurement}

\item{dateCol}{string of the column name containing the date information. This can be either a recognized date object (e.g., Date, POSIX)}

\item{repCol}{string of the column name containing the replicate information}

\item{...}{additional arguments passed to function}

\item{infoCols}{integer vector; Any columns in the sizeInfo object that are not the taxonomic ID, sampling metadata, or size class columns}
}
\description{
This is the main function of the secpRod package. It will calculate secondary production for all groups based on the methods described in the taxa information object. Depending on input values varying summaries are returned.
}
