% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_production.R
\name{calc_production}
\alias{calc_production}
\title{calc_production}
\usage{
calc_production(
  taxaSampleListMass = NULL,
  infoCols = NULL,
  taxaInfo = NULL,
  bootNum = 100,
  wrap = 1L,
  taxaSummary = "full",
  ...
)
}
\arguments{
\item{taxaSampleListMass}{data.frame of sample length-masses and abundances}

\item{infoCols}{integer vector; Any columns in the sizeInfo object that are not the taxonomic ID, sampling metadata, or size class columns}

\item{taxaInfo}{dataframe (or coercible); The taxonomic information for calculating secondary production. This must include a taxonomic ID column with the same name as that of \code{taxaSampleListMass}}

\item{bootNum}{integer. The number of bootsrapt samples to create}

\item{wrap}{logical should an extra date be added to make a full calendar year?}

\item{taxaSummary}{string of \code{'short'}, \code{'full'}, or \code{'none'} to distinguish the information returned}

\item{...}{additional arguments passed to function}
}
\description{
This is the main function of the secpRod package. It will calculate secondary production for all groups based on the methods described in the taxa information object. Depending on input values varying summaries are returned.
}
