% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_prod.sample.R
\name{igr_prod.sample}
\alias{igr_prod.sample}
\title{igr_prod.sample}
\usage{
igr_prod.sample(
  df = NULL,
  dateDf = NULL,
  lengthValue = NULL,
  massValue = "mass",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  wrap = FALSE,
  full = TRUE,
  ...
)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{lengthValue}{string of the column name containing the length class measurements}

\item{massValue}{character string identifying the column name of the mass value}

\item{abunValue}{character string identifying the column name of the density value}

\item{dateCol}{character string identifying the column name of the sample date information. This is distinguished from *Value parameters in that Values may be used to maintain units provenance in the future. This may also change.}

\item{repCol}{character string identifying the column name of the replicate information.}

\item{wrap}{logical. should the calculations be wrapped by adding an additional date to make a full year?}

\item{full}{logical. should the full summary be returned with mean and sd.}

\item{...}{additional arguments passed to function, including variables to predict growth rate from growth function}

\item{sizesDf}{a data.frame of the sizes}
}
\value{
list object with annual production, mean biomass, and mean abundance of the sample
}
\description{
This function calculates taxa production based on the instantaneous growth method
}
