% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prod.sample.R
\name{is_prod.sample}
\alias{is_prod.sample}
\title{is_prod.sample}
\usage{
is_prod.sample(
  df = NULL,
  dateDf = dateDf,
  massValue = "afdm_mg",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  wrap = FALSE,
  full = TRUE,
  ...
)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{dateDf}{data frame of date information sampling date such as interval length in days.}

\item{massValue}{string of the column name containing the mass measurement}

\item{abunValue}{string of the column name containing the abundance or density measurement}

\item{dateCol}{string of the column name containing the date information. This can be either a recognized date object (e.g., Date, POSIX) or numeric.}

\item{repCol}{string of the column name containing the replicate information}

\item{full}{logical. should the summary be returned with mean and sd}

\item{...}{additional arguments passed to function}
}
\value{
list object with annual production, mean biomass, and mean abundance of the sample
}
\description{
This function calculates taxa production based on the increment-summation method
}
