% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_prod.sample.R
\name{pb_prod.sample}
\alias{pb_prod.sample}
\title{pb_prod.sample}
\usage{
pb_prod.sample(
  df = NULL,
  massValue = massValue,
  abunValue = abunValue,
  dateCol = dateCol,
  repCol = repCol,
  pb = NULL,
  wrap = FALSE,
  full = TRUE,
  ...
)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{massValue}{character string identifying the column name of the mass value}

\item{abunValue}{character string identifying the column name of the abundance value}

\item{dateCol}{character string identifying the column name of the date column}

\item{repCol}{character string identifying the column name of the replicate column}

\item{pb}{numeric. The production:biomass ratio.}

\item{wrap}{logical. should the calculations be wrapped by adding an additional date to make a full year?}

\item{full}{logical. should the full summary be returned with mean and sd}

\item{...}{additional arguments passed to function}
}
\value{
list object with annual production, mean biomass, and mean abundance of the sample
}
\description{
This function calculates sample production based on the production:biomass method
}
