% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_prod.sample.R
\name{sf_prod.sample}
\alias{sf_prod.sample}
\title{sf_prod.sample}
\usage{
sf_prod.sample(
  df = NULL,
  sizesDf = NULL,
  massValue = NULL,
  massLabel = NULL,
  cpi = NULL,
  full = TRUE,
  ...
)
}
\arguments{
\item{df}{a data.frame of long format returned from \code{convert_length_to_mass()} function}

\item{sizesDf}{a data.frame of the}

\item{massValue}{character string identifying the column name of the mass value}

\item{massLabel}{character string identifying the column name of the mass value}

\item{cpi}{integer. The cohort production interval.}

\item{full}{logical. should the full summary be returned with mean and sd}

\item{...}{additional arguments passed to function}
}
\value{
list object with taxa summary of the sampled data
}
\description{
This function calculates taxa production based on the size-frequency method
}
